Reinstated full Torch support. Pytest now supported.
- The new `symtensors/tests/` directory contains an exhaustive list of
  auto-updated unit tests for the SymmetricTensor API.
- Each tensor format/backend gets its own test module/page.
- Tests are by default all sourced from `symtensors/testing/api.py`.
  Thus they don’t need to be rewritten for each class, unless they
  need to be changed.
  Also, if an API test changes, it is reflected in the test for all
  subclasses (unless it was overridden). Hence tests are “auto-updated”.
- The extra content provided to run tests in a Jupyter Book makes it
  very convenient to browse, test and modify tests.

Currently the tests only cover the DenseSymmetricTensors.
Implementing tests for other formats should be just a matter of
copying files and changing the few lines at the top that specify the
SymmetricTensor subclass to test.

In getting all the tests to run for the TorchDenseSymmetricTensor, a
lot of fixes were added to ensure full Torch compatibility. In particular,
the family of `TorchSymmetricTensor`s now provides implementations for
the same functions as `SymmetricTensor`, so algorithms using `np.multiply`
or `utils.symmetrize` work transparently with both backends.